server/requirements.txt
python-docx
Flask 
flask-cors